Header component is now used globally
