+++ sudokuru/app/Pages/Lessons/AdjustNotesLesson.tsx
@@ -10,7 +10,7 @@ const AdjustNotesLesson = () => {
 
     return (
         <View>
-            <Header></Header>
+            <Header page={'Lesson'}/>
             <View style={homeScreenStyles.home}>
                 <View style={homeScreenStyles.homeMenu}>
                     <SidebarMenu></SidebarMenu>
