+++ sudokuru/app/Pages/Lessons/BoxLineReductionLesson.tsx
@@ -10,7 +10,7 @@ const BoxLineReductionLesson = () => {
 
     return (
         <View>
-            <Header></Header>
+            <Header page={'Lesson'}/>
             <View style={homeScreenStyles.home}>
                 <View style={homeScreenStyles.homeMenu}>
                     <SidebarMenu></SidebarMenu>
