+++ sudokuru/app/Pages/Lessons/HiddenSetLesson.tsx
@@ -10,7 +10,7 @@ const HiddenSetLesson = () => {
 
     return (
         <View>
-            <Header></Header>
+            <Header page={'Lesson'}/>
             <View style={homeScreenStyles.home}>
                 <View style={homeScreenStyles.homeMenu}>
                     <SidebarMenu></SidebarMenu>
