+++ sudokuru/app/Pages/Lessons/HiddenSingleLesson.tsx
@@ -10,7 +10,7 @@ const HiddenSingleLesson = () => {
 
     return (
         <View>
-            <Header></Header>
+            <Header page={'Lesson'}/>
             <View style={homeScreenStyles.home}>
                 <View style={homeScreenStyles.homeMenu}>
                     <SidebarMenu></SidebarMenu>
