+++ sudokuru/app/Pages/Lessons/NakedSetLesson.tsx
@@ -10,7 +10,7 @@ const NakedSetLesson = () => {
 
     return (
         <View>
-            <Header></Header>
+            <Header page={'Lesson'}/>
             <View style={homeScreenStyles.home}>
                 <View style={homeScreenStyles.homeMenu}>
                     <SidebarMenu></SidebarMenu>
