+++ sudokuru/app/Pages/Lessons/NakedSingleLesson.tsx
@@ -10,7 +10,7 @@ const NakedSingleLesson = () => {
 
     return (
         <View>
-            <Header></Header>
+            <Header page={'Lesson'}/>
             <View style={homeScreenStyles.home}>
                 <View style={homeScreenStyles.homeMenu}>
                     <SidebarMenu></SidebarMenu>
