+++ sudokuru/app/Pages/Lessons/SimplifyNotesLesson.tsx
@@ -10,7 +10,7 @@ const SimplifyNotesLesson = () => {
 
     return (
         <View>
-            <Header></Header>
+            <Header page={'Lesson'}/>
             <View style={homeScreenStyles.home}>
                 <View style={homeScreenStyles.homeMenu}>
                     <SidebarMenu></SidebarMenu>
