+++ sudokuru/app/Pages/Lessons/SinglesChainingLesson.tsx
@@ -10,7 +10,7 @@ const SinglesChainingLesson = () => {
 
     return (
         <View>
-            <Header></Header>
+            <Header page={'Lesson'}/>
             <View style={homeScreenStyles.home}>
                 <View style={homeScreenStyles.homeMenu}>
                     <SidebarMenu></SidebarMenu>
