+++ sudokuru/app/Pages/Lessons/SwordfishLesson.tsx
@@ -10,7 +10,7 @@ const SwordfishLesson = () => {
 
     return (
         <View>
-            <Header></Header>
+            <Header page={'Lesson'}/>
             <View style={homeScreenStyles.home}>
                 <View style={homeScreenStyles.homeMenu}>
                     <SidebarMenu></SidebarMenu>
