+++ sudokuru/app/Pages/Lessons/XWingLesson.tsx
@@ -10,7 +10,7 @@ const XWingLesson = () => {
 
     return (
         <View>
-            <Header></Header>
+            <Header page={'Lesson'}/>
             <View style={homeScreenStyles.home}>
                 <View style={homeScreenStyles.homeMenu}>
                     <SidebarMenu></SidebarMenu>
